Add a Context class to simplify ProcessClientHello
This CL creates a new class - ProcessClientHelloContext - which holds all of the arguments passed to ProcessClientHello, and carries them along through asynchronous processing.  This allows the argument lists of all methods involved to be drastically reduced.  It also naturally takes over the done-callback-invoking duties of the ProcessClientHelloHelper.

gfe-relnote: No-op refactoring in QUIC ClientHello processing, not flag-protected.
PiperOrigin-RevId: 239045806
Change-Id: I353f9880fded792c61757309d41addbd902403f8