Create a RandomDelayLink (and symmetric version) and use it in QuartcBidiTest.
Random delay is one of the features of the network simulation Quartc uses in its
upper-layer quality tests.  It is used to simulate poor network profiles where
random interference or delays may occur.

This change extends OneWayLink to allow tests to specify a random delay (to be
added to the transfer time and propagation delay).  The additional random delay
follows an exponential distribution, with the tail chopped off at the chosen
maximum delay.

QuartcBidiTest adds a case which uses random delay with a median of 25 ms.

gfe-relnote: n/a (test only)
PiperOrigin-RevId: 244067695
Change-Id: I83c05cc5d2b9973644b501f384ed3c431ae761fd