Make QUIC EXPECT and ASSERT test macros reject comparison of unsigned integers
This CL overrides the CHECK/DCHECK/EXPECT/ASSERT macros. Unfortunately these macros allow comparison of integers of different signs, which allows us to write EXPECT_EQ(a, b) when a and b have different signs. However, the equivalent macros in Chromium do not allow this mismatch, and this has caused as number of quiche merge issues. Therefore, to avoid those, we add a non-executed comparison to the macros which ensures that compilation fails if there is a comparison of different signs. An example of the fix previously required for quiche merge can be found in cl/244853842.

gfe-relnote: n/a, changes only impact compile-time for debug builds and tests
PiperOrigin-RevId: 245325129
Change-Id: I1e4d0bb424754c62e62a72378789ae6902eb9af5