Create a helper that wraps a competing QuicEndpoint for Quartc simulator tests.
This helper owns a QuicEndpoint and a QuicAlarm, and schedules sends of X bytes
every Y ms on that endpoint.  This is the basic operation we want from competing
flows, and simplifies test cases that use competing flows by removing all the
boilerplate of setting up and maintaining alarms.

gfe-relnote: n/a (quartc test only)
PiperOrigin-RevId: 246907891
Change-Id: I23d9b65d5c97c3e5863f7f323935dae9f55a2a3a