Fix QpackRoundTripFuzzer when cookies are split along '\0'.
The root of the issue is that ValueSplittingHeaderList splits the value of the
cookie header along '\0', but SpdyHeaderBlock::AppendValueOrAddHeader() joins
them with "; " separators.  The long term solution will be changing
ValueSplittingHeaderList to split cookies along semicolons, but this logic in
QpackRoundTripFuzzer will still be necessary, becaue "value1;value2" will be
transformed by an encoding-decoding round-trip to "value1; value2", which
QpackRoundTripFuzzer should not crash on.

Also include regular and minimized test cases from b/131395084, b/131409242, and
https://crbug.com/957840.

gfe-relnote: n/a.  Change in fuzzer only.
PiperOrigin-RevId: 247281474
Change-Id: I405066d45ae5e6cbf4c53d8a4f3afb1d0f6b3327