gfe-relnote: Construct QuicMemSliceSpans from a QuicMemSlices without copies.
QuartcSession uses this functionality to send queued message frames without
copying the data backed by QuicMemSlices in its queue.

Note that this extends the platform/api by adding a new constructor to
QuicMemSliceSpan.  It adds a corresponding constructor to QuicMemSliceSpanImpl,
so this will need to be merged into chromium manually.  The changes required
should be minimal (<10 lines of code).

Not flag-protected, since the only changes are (a) platform/ code that is not
called anywhere except quartc/ and (b) quartc/ code that is not called in GFE.

PiperOrigin-RevId: 247689071
Change-Id: Ifbfbb52294b41da45a230ab63f8f199754b68211