Fix threading in EndToEndTestWithTls.NoUndecryptablePackets
TSAN detected a threading issue in this test where the server stats are accessed from the main thread while the server thread is still running (and can write to them while we're reading). This CL matches the convention in the rest of the file which is to pause the server thread before accessing the server stats and resume it afterward.

gfe-relnote: n/a, test-only change
PiperOrigin-RevId: 250244377
Change-Id: I167416683b9030df275ef73062974f64819fbb0d