Propagate Datagram ACKs from QuartcSession
Introduce maps of message_id to datagram ids, so we could translate message ACKs received from QUIC to datagram ACKs that are propagated up the stack. We have to maintain map, because QUIC assigns message_id only when message is sent (it can be queued if connection was congestion controlled).

WebRTC will use transport sequence numbers as datagram_id so that ACKs can be mapped to sent packets.

gfe-relnote: n/a (Quartc only)
PiperOrigin-RevId: 250589798
Change-Id: Ib7df728873a839bd3343adb182fb68c3bf055caf