gfe-relnote: Pass receive_timestamp to OnMessageAcked().
When receive timestamps are turned on and message frames are enabled, this will
make receive timestamps available to the session whenever a message is acked.

This will be used for RTP over QUIC, where ack timestamps are required for RTP's
congestion controller and media adaptation.  (The current implementation uses
QUIC like a tunnel and disables its congestion control.)

Not flag-protected: no change in behavior, just a change in what is visible
where.

PiperOrigin-RevId: 250703980
Change-Id: I01bcee29d5dfb808771c54e4fd14ebbfa8750375