Force quic_optimize_inflight_check off for Quartc.
This optimization is apparently incompatible with datagram/message-only flows.
The optimization circumvents a check to see if there are any retransmittable
frames, relying on a previous check for any bytes_in_flight instead.  Message
frames count toward bytes_in_flight, but are not retransmittable.

When this is enabled, some quartc simulator tests can land in an infinite loop
of retransmission timeouts.

gfe-relnote: n/a (Quartc only)
PiperOrigin-RevId: 250772361
Change-Id: I338c9c8958538874c77b74f3bcf4f8b9bd7e74e9