Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Trying to find a way to add a reviewer...
> 
> You should be able to just add it like in regular Chromium gerrit.

No, it fails to autocomplete any reviewers (and, if I explicitely add any, they say the user is not missing). I guess I may be missing some permission, so it does not allow me to refer to actual users?

> 
> FWIW, I believe this was already fixed in https://quiche.googlesource.com/quiche/+/014f56d6a10b0f4806d13e109dc9a1e933ec19fb

For the record, it is this one. And yes, that's including the fix for the GCC issue. Thanks!
https://quiche.googlesource.com/quiche/+/9424add9d73432a794b7944790253213cce6dcb8

Patch-set: 1
Reviewer: Gerrit User 1000083 <1000083@6cd04770-3c6a-3732-a6a3-94aeb48175de>