Additional functional version tests.
This CL changes some version-99 tests to be functional, based on
methods added to quic_versions.h. Instead of things like "if version==123"
there are "if version_supports_feature_x(version)". Currently, the method
tests for the correct, existing, versions (that is, if it tested for
version 99 before, it still does.)

gfe-relnote: N/A. This CL changes version-99 tests to be functional.
PiperOrigin-RevId: 252875457
Change-Id: I68b5f1bbd9513a7abd689af88e3c18c67e168ded