Make QpackEncoder interface not progressive.
When the QPACK implementation was originally written, an
HpackEncoder::ProgressiveEncoder interface was used for encoding.  However, the
way it is used in QuicSpdyStream, it is clear now that encoding can be done in a
single pass into a std::string object.  QuicStream::WriteOrBufferData() does not
lend itself easily to notifying the QPACK encoder that it is ready to take more
data.  Instead it is more natural to encode the entire header block and buffer
that in QuicStreamSendBuffer.  Also note that this should not cost more memory
than doing progressive encoding, because the SpdyHeaderBlock object would need
to be kept around if progressive encoding was used.

This CL introduces no functional change, it just changes the interface, and
moves the code that calls Next() from QuicSpdyStream into QpackEncoder.  A
follow-up CL will clean up QpackEncoder internals.

gfe-relnote: n/a, QUIC v99-only change.
PiperOrigin-RevId: 254323435
Change-Id: Ic3d203ed9f019ad01dba61d407a6e73bde5d12d1