Alter QuicConnectionDebugVisitor::OnConnectionClosed to take QuicConnectionCloseFrame
This continues the work starting in CL/254029189, extending the API change to QuicConnectionDebugVisitor::OnConnectionClose. It takes a QuicConnectionCloseFrame with all the information about the close instead of then explicit parameters from the frame.

This is done as part of the work to support the IETF QUIC Connection Close frame.

gfe-relnote: N/A - just changes method signature, logic remains the same.
This is not flag protected.
PiperOrigin-RevId: 254744908
Change-Id: Id80be9b2fa65d6cf8b411ec322074cf4aa7c6bae