Do not handle impossible outcomes in HttpDecoder.
At a number of places HttpDecoder has DCHECKs in place that guarantee the input
to QuicDataReader methods is valid, yet it handles failure.  This CL replaces
those by DCHECKs.  Not only is this shorter and more sane, it also cuts down on
debug strings.

gfe-relnote: n/a, QUIC v99-only change.
PiperOrigin-RevId: 254808999
Change-Id: Ic8750f58352153f40f0adc306792612b13ee616f