Fix QUIC version used when computing ALPN
The ALPN should be based on the QUIC version that is currently in use, not the first one we tried prior to version negotiation. Running the newly added version negotiation end-to-end test with TLS enabled is what found the issue.

gfe-relnote: fix ALPN, protected by flag quic_supports_tls_handshake
PiperOrigin-RevId: 257666341
Change-Id: I4f1619cf4923fae01443ed4c89450df0be78cd39