Add retire_prior_to to NewConnectionId frame
IETF altered the NewConnectionId frame format, adding a "retire_prior_to"
field. This CL adds that field and modifies the framer and tests as needed.

gfe-relnote: N/A is only for IETF QUIC (version 99 flag protected)

NEW_REQUIRED_FIELD_OK="Message not currently in use, nothing in the logs"

Note wrt the protobuf update - this CL adds a field to an IETF QUIC frame. Correspondingly, the relevant protobuf gets a new field. Since IETF QUIC is not in use, there are no instances of this protobuf in use/existance/etc yet ... so there is no incompatbility with other processes/etc to be concerned with.

PiperOrigin-RevId: 257833679
Change-Id: I80d80db9f1f23b8e570c504199fdd346b7cb4ca4