Update IETF QUIC Stop Sending to use varint for error code
Recent update to the IETF QUIC specification changed the serialization of
the Stop Sending frame's error code from uint16_t to varint ... this CL
implements that change.

Note that we still keep the error code as a uint16_t -- a later CL will update
the internal storage of error codes to uint64_t.

gfe-relnote: N/A change to IETF QUIC (V99) only.
PiperOrigin-RevId: 258404381
Change-Id: I4872cd5fce68454978880a8009c2de7ada43bdf8