Automated g4 rollback of changelist 258400699.
*** Reason for rollback ***

Chromium does not support dynamic_cast, so this cannot be merged as is.  It will take some time to figure out how to determine if session() is a QuicSpdyClientSession instance or not, so I need to revert it for the time being to unblock the QUICHE merge.

*** Original change description ***

gfe-relnote: Add support for sending MAX_PUSH_ID, defaulting to zero and close connection if we receive a push ID higher than the max.
Protected by the existing disabled and blocked quic_enable_version_99 reloadable flag.

***

PiperOrigin-RevId: 258464901
Change-Id: Ic738f37737850cff8f46a30491637b3314c99144