gfe-relnote: In QuicSpdyClientBase, automatically convert HTTP request header names to lower case. Protected by default true, non-feature flag --quic_client_convert_http_header_name_to_lowercase.
This code should not be used in production GFEs. The only place where QuicClient is used AFAIK is in L0's udp proxy, which only connects to L1s and never sends HTTP requests.

Tested with a quic_client sending a request with --headers=X-Return-Encrypted-Headers:all, which succeeded(200) when --quic_client_convert_http_header_name_to_lowercase is true, failed(400) when the flag is false.

PiperOrigin-RevId: 258548285
Change-Id: I04ebd4041b9da3c80a0f5ff8b30b20627e02caf1