Call MarkConsumed() only once in QuicSpdyStreamBodyBuffer::MarkBodyConsumed() with a combined byte count.
This is a small behavioral change in preparation for a future CL which will make
MarkBodyConsumed() and also ReadBody() return the number of bytes to be consumed
by the caller.  Splitting out this change allows that future CL to be a pure API
change.  The ultimate motivation is much cleaner tests for
QuicSpdyStreamBodyBuffer.

gfe-relnote: n/a, change in QUIC v99-only code.
PiperOrigin-RevId: 259424818
Change-Id: I1a96285ac7fcdfc9eae2768c9f2b64b82a323929