Update QuicSpdySessionTestClient.Http3ServerPush* tests.
"\x1header" is not a valid input for a unidirectional stream: a push stream must
have a Push ID immediately following the stream type varint.

My motivation of writing this CL is to be able to land cr/258682203.  That change
breaks these test, because "header" is parsed as the beginning of a frame of
unknown type and some of it gets consumed.  (Note that currently the push ID
is not parsed due to a bug.)

gfe-relnote: n/a, test-only change.
PiperOrigin-RevId: 260311915
Change-Id: If4f84a811d51f768af06421e2600692af059bd2e