Send correct stateless reset token when using TLS
When using TLS, the stateless reset token needs to be populated in the QuicConfig early in the session's lifetime to ensure it's correctly encoded in the server's transport parameters. Previously the server would send an all-zeroes token. I've confirmed that EndToEndTestWithTls.SendStatelessResetTokenInShlo fails without the fix and passes with the fix.

gfe-relnote: send correct reset token, protected by disabled TLS flag
PiperOrigin-RevId: 261374750
Change-Id: I60725c5eb6cedfa1b06cf3797bf258875fb495eb