Add VersionNegotiationProbeEndToEnd test
This CL introduces SavingWriter, a subclass of QuicPacketWriterWrapper that saves all the packets instead of sending them. It uses it to test that a packet generated by QuicFramer::WriteClientVersionNegotiationProbePacket does generate a response and that the response can be correctly parsed by QuicFramer::ParseServerVersionNegotiationProbeResponse.

This CL also fixes a typo in a log statement.

gfe-relnote: n/a, log-only and test-only change.
PiperOrigin-RevId: 263387682
Change-Id: I2dd2eccd5db973b93f0aa8e56c155b47e9b50855