Add flag to QuicSession indicating that it has been configured
This CL adds a flag to QuicSession that indicates when the QuicSession has been
configured. This needed for follow-on work where IETF QUIC is limited as to the
frames that can be sent _until_ the transport config has been performed.

QuicSession test has been modified to A) artificially perform a configuration
and B) check at test termination that it has been configured.

gfe-relnote: N/A not significant. Other changes all to tests.
PiperOrigin-RevId: 264362154
Change-Id: I1467cd4ed4531c01950d3fb938637e8633de815e